apps/api-server-proxy/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9 
gevent==24.11.1
python-dotenv
pezzo